# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ArchiveMaster API",
        "description": "Sistema de gestión documental empresarial completo",
        "contact": {
            "name": "ArchiveMaster Support",
            "email": "admin@archivemaster.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "ArchiveMaster API Server"
        }
    ],
    "paths": {
...
        "/api/webhooks/register": {
            "post": {
                "tags": [
                    "Webhooks"
                ],
                "summary": "Registrar webhook",
                "description": "Registrar un nuevo webhook para recibir notificaciones de eventos",
                "operationId": "3327efb80281af7bb8a3e629b029cf47",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "url",
                                    "events"
                                ],
                                "properties": {
                                    "url": {
                                        "description": "URL del endpoint webhook",
                                        "type": "string",
                                        "format": "url",
                                        "example": "https://mi-sistema.com/webhooks/archivemaster"
                                    },
                                    "events": {
                                        "description": "Eventos a los que suscribirse",
                                        "type": "array",
                                        "items": {
                                            "type": "string"
                                        },
                                        "example": [
                                            "document.created",
                                            "document.updated",
                                            "document.status_changed"
                                        ]
                                    },
                                    "name": {
                                        "description": "Nombre descriptivo del webhook",
                                        "type": "string",
                                        "example": "Sistema ERP Principal"
                                    },
                                    "secret": {
                                        "description": "Secreto para validar la autenticidad de las llamadas",
                                        "type": "string",
                                        "example": "mi_secreto_super_seguro"
                                    },
                                    "active": {
                                        "description": "Si el webhook está activo",
                                        "type": "boolean",
                                        "example": true
                                    },
                                    "retry_attempts": {
                                        "description": "Número de reintentos en caso de fallo",
                                        "type": "integer",
                                        "example": 3
                                    },
                                    "timeout": {
                                        "description": "Timeout en segundos para las llamadas",
                                        "type": "integer",
                                        "example": 30
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Webhook registrado exitosamente",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Webhook registrado exitosamente"
                                        },
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "string",
                                                    "example": "webhook_123456"
                                                },
                                                "url": {
                                                    "type": "string",
                                                    "example": "https://mi-sistema.com/webhooks/archivemaster"
                                                },
                                                "events": {
                                                    "type": "array",
                                                    "items": {
                                                        "type": "string"
                                                    }
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Sistema ERP Principal"
                                                },
                                                "active": {
                                                    "type": "boolean",
                                                    "example": true
                                                },
                                                "created_at": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Errores de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Errores de validación"
                                        },
                                        "errors": {
                                            "type": "object"
                                        },
                                        "timestamp": {
                                            "type": "string",
                                            "format": "date-time"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Usar Bearer Token para autenticación",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}